SVG import: allow user to set custom dimensions
With this, I consider SVG import a finished feature!

I've debated exposing some low-level import options in the SVG import screen (like allowing the user to specify speed vs quality trade-offs - right now everything defaults to "max quality"), but that introduces a lot of localization work for a fairly esoteric set of parameters.  As such, I'm going to postpone the idea for now, but I'm willing to revisit if anyone finds that kind of granular control useful.