Add "Map" category to Adjustments menu
This parent menu will hold at least:

> Gradient map...
> Palette map...

Palette mapping is already implemented via the Effects > Stylize > Palettize menu, but that is a non-intuitive place to look.  (And it doesn't appear in searches for "palette" because of how it's spelled, which is doubly bad!)  So I've simply mirrored that dialog via the new Palette Map menu.

Gradient mapping is one of only a few adjustments that Photoshop supports but PD does not, so I am very excited to figure it out!  (It is also a blocker for adjustment layers, since that's its primary usage in PS - so let's get it implemented here in a smart, adjustment-layer-friendly way.)

Actual gradient map code is still TODO lol