Improve dropdown width on non-en-US locales
Many thanks to Roy K for pointing out this issue to me.  (And thank you to him for an updated German localization.)

Dropdown controls will now do a better job of widening the dropdown itself when localized text is larger than the size of the base control.  This should make e.g. the Blend Mode dropdown (in the layer toolbox) much more useable under certain locales.