pdTitle: improve layout behavior at tiny sizes
The left-side toolbox is an interesting pain point in PD's layout.  The biggest issue is localized text, which is *incredibly* variable in its length relative to the original English text.

In the past, if text couldn't fit, PD would simply shrink the text as small as it could then just let it overflow everywhere.  We can do better!

Now PD will add ellipsis and manually clip the text.  I've also reworked the arrow design to look better, specifically improving its positioning (aligned against the font's x-height) and making it nearly identical to the arrow used by dropdowns.  This looks significantly better in the new toolpanel design as well (which uses a *lot* of titlebars).