Localization engine: start rolling out a new hash-table strategy...
...for accelerating translated text lookup.

In this initial commit, the fast dictionary lookup is not persistent across sessions.  It is generated on-the-fly as translations are queried.

When a new phrase is encountered (one that requires localization), we first query a fast hash-table of all translations already retrieved this session.  Only on table misses do we travel all the way out to the base XML localization file and pull the translation from there (after which, we add it to the fast table).

This approach provides significant time savings vs the original strategy, but that's not really the point as all translation times combined take up only fractions of a second during a typical session.  (PD uses a custom-built Boyer-Moore search strategy for translation retrieval, which is extremely fast vs traditional InStr/StrStrW approaches.)

What's more meaningful is that this will provide a way for me to support new translation features that would otherwise be untenable performance-wise - like allowing translators to supply per-control-instance text, for example, if I've reused English text across two control instances but that reuse doesn't work in another language.  (Gendered versions of e.g. adjectives are especially problematic this way.)

I'm going to roll out further translation engine changes slowly, because these are very difficult for me to test in a comprehensive way and I don't want to break anything.  (I've tried to test this commit as exhaustively as I can, while not speaking the various languages I can use for testing lol.)  Please let me know if you encounter any weirdness with non-en-US versions of the program.

Thank you also to Roy K for an updated German localization