PSD import: add support for retrieving layer group masks
Relates to #391.  Thank you to @ImportTaste for filing the original bug report.

This is the first of several commits aimed at successfully round-tripping layer group masks from PSD files.  In this first commit, PhotoDemon can now successfully retrieve group masks, and if a debug constant is enabled, it'll even load them as the pixel data for the "dummy" layer group layers we create.  (This obviously causes the composite image to display weirdly, but it's very helpful to me so I can see what's happening with a given group.)

My next task is to "tag" group layers with mask data (as an optional add-on) and ensure that data persists for the life of the target group.  The final task will then be to figure out how to export the mask in a PSD-compatible way.