Fix compositing of affine-transformed layers with...
...a combination of weird criteria:

1) non-destructive resize active
2) no non-destructive rotation or skew active
3) positioning beyond the left/bottom corner of the parent image
4) a blend-mode other than "normal"

This extremely specific set of criteria was triggering an unexpected branch that would cause the target layer to be composited to [0, bottom-aligned] on the final image.  This case is now fixed.

(Bugs like this make me weirdly proud, since they're a testament to how complex and overly optimized PD's rendering chain has become.)