Improve mechanism for triggering "flatten" dialog
"Flatten" behavior is inconsistent across popular photo editors.

Some always "flatten" the image to 24-bpp, without an alpha channel.

Some "flatten" to 32-bpp if alpha is present in any one layer.

Others "flatten" to 32-bpp IFF the composited image exposes alpha.

PD tries to optimize this case and not bother the user with unnecessary prompts.  If the composited image will be opaque, PD skips the prompt.  If the composited image contain transparency, PD displays a prompt with the option to keep or remove transparency from the flattened image (with choice of matte color, as appropriate).