XCF importer: add support for most HDR formats
All HDR floating-point and integer formats are now working, except for half-precision floats which are producing strange results.  I suspect that GIMP is encoding these slightly differently than LittleCMS is interpreting them, but I'll need to write manual half-to-float conversion routines to inspect further.  (This is also a minor use-case, since 16-bit integer HDR is way more prevalent.)

The code for all these formats is extremely ugly, but it will be cleaned up soon.  Reverse-engineering minutiae of each encoding scheme is non-trivial so a lot of experimenting was required to arrive at correct results!