Image resampling: nice performance boost on integer mode
By using a fixed power-of-two for integer estimation, VB6's compiler can be "coerced" into generating bit-shifts instead of integer divides on the sampling inner loop.  This provides a nice little performance boost in "fast" resampling mode, at a minor hit to quality relative to the old version.  (Note that a "high-quality" resampling mode is still available in the UI, which uses floating-point math for everything.)