XCF import: finish layer mask support (all color-depths)
Masks are encoded at the same color-depth as the rest of the image, so e.g. 32-bit float channels have 32-bit float masks.   This complicates their handling.

PhotoDemon now correctly handles layer masks in all precisions, both int and float, and each of indexed and grayscale and RGB/A modes.

At present, masks are auto-merged into each imported layer's alpha channel, but I've designed the importer to make it *very* simple to switch to a dedicated layer mask class in the future!