Printing: clean up code and remove lingering picture boxes
PhotoDemon's printer support is extremely meh.  To be honest, I'm not highly motivated to improve it, but this commit at least migrates all picture boxes to pdPictureBox (for the XP print dialog) and improves printing performance for the Vista+ print interface by using uncompressed PNGs for interop.  I've also reworked the code in the XP print dialog to be less of a horror show, although it's still nowhere near the quality I usually try to achieve in PD.  Oh well.

To be honest, I may end up finally "fixing" this code as part of a future export to PDF option (I've gotten requests for this for multipage TIFF files, with FastStone Image Viewer given as a reference - studying this is TBD).  Until then, please use other software for printing!