Effects > Transform > Perspective: show live preview...
...of transformed image.  The new preview feature uses a stripped-down version of PhotoDemon's full projective transform code, with a number of preview-specific tweaks to hit a sweet spot between quality and performance.  I'm especially pleased with the preview's handling of a background checkerboard on semi-transparent images - the checkerboard actually behaves correctly!

As part of this work, a function was needed to determine if the target quadrilateral is convex or not.  (If it isn't convex, we want to suspend rendering in the preview area to prevent visual glitches.)  Thank you to Rory Dalton for an optimized algorithm he shared at StackOverflow:

https://stackoverflow.com/questions/471962/how-do-i-efficiently-determine-if-a-polygon-is-convex-non-convex-or-complex

I altered the algorithm slightly to cover some specific usage points in the way PD interacts with it, so please blame me for any resulting inconsistencies in the code.