Effects > Transform > Perspective: more UI improvements
- Spinner controls are now available for all corner coordinates, allowing for extremely fine-grained control over the transform.
- I've reworked the way the preview area is calculated to make better use of the available interactive space

It's nice to revisit old tools sometimes and clean them up a bit.  This tool is in a much better place now!

At present, I think the transform is actually slightly off due to an off-by-one error in the way dimensions are calculated, but I need to double-check my math before modifying this - final solution TBD!