pdInpaint: a new solution for content-aware fill
Photoshop has a borderline-magical feature called "content-aware fill":

https://helpx.adobe.com/photoshop/using/content-aware-fill.html

I want something similar for PhotoDemon, so here we gooooooo!

The best reference I've seen on texture synthesis (which is what content-aware fill does - it attempts to "synthesize" a new "texture" that can be used to seamlessly "fill" over the top of an unwanted object) is Paul Harrison's PhD Thesis, "Image Texture Tools: Texture Synthesis, Texture Transfer, and Plausible Restoration":

https://www.logarithmic.net/pfh-files/thesis/dissertation.pdf

Using the concepts from his thesis, Paul would go on to release the "Resynthesize" plugin that provides content-aware fill in GIMP.  Resynthesize is still available as a GIMP plugin, but unfortunately it is under a new maintainer now and it is very challenging to build for Windows (and the source code is GPL anyway, so I can't use it in PhotoDemon even if I wanted to).

Fortunately, Paul's thesis is very comprehensive.  He lays out pretty much everything you need to consider when writing a texture synthesizer, and he cites many other author's takes on the topic (which are also useful references, with good ideas to be taken from all).

From these various references, I have now completed my first draft of a texture synthesizer for PhotoDemon.  This first draft is not actually a content-aware fill tool - that will come later - but it *is* a seamless tile generator, which provides a simple test ground for the underlying algorithm.  Basically, given an input image, the code tries to construct a novel output image that is capable of seamless tiling, using only pixel data from the original image as its reference.  From this, it's easy to study how the algorithm handles complexity and sampling, and whether it's capable of generating representative textures from a variety of inputs.

As it stands today, this code is slow, but it works well - very well, even - and I'm now at a stage where I comfortable committing it so I have a good backup reference before I start writing a UI for it.  (A UI is needed for further testing, since there are a number of parameters you can toggle to trade-off between quality and performance, and I need to ensure PD's algorithm works well across the entire spectrum.)

I don't have an ETA for final merge into PD, but it shouldn't be long - I'm thinking a measurement in days possibly weeks, not months.