First draft of `Edit > Content-Aware Fill` tool
Holy shit, friends - this actually works.  I'm honestly a little (okay a LOT) astonished to see the functional content-aware fill behavior in action, because part of me thought this was never going to work in a reasonably good way.

But it works, and it works *well*.  Like, this will be a tool that people actually use!

There's a ton of clean-up and refinement left to do (and I still need to build a UI) but that's easy-peasy compared to just getting the damn algorithm working in the first place.