Content-aware fill: large performance boost
Like pretty much all algorithms, content-aware fill can be designed against a variety of performance vs memory trade-offs (e.g. you can use a lot more memory to gain a lot more speed).

My first draft didn't really account for this, so in some places it used minimal memory while in others it used a ton but was very fast.  Time to start aligning these goals!

In this commit, the in-painter now requires the source, destination, and mask surfaces to all be the same size.  (Previously, the mask could be any arbitrary size relative to the src/dst images; the inpainter would handle custom bounds-checking accordingly.)

The main benefit of this is that it hugely simplifies boundary checking on pixel accesses.  We can perform all boundary checks up-front and then use a single set of compares across *all* subsequent access checks, which provides a ~30% performance improvement.  (For example, a fill that used to take 6 seconds now takes 4.)

To combat the memory increase required by a larger mask, the parent function now pre-crops out only the relevant region of the full image prior to in-painting.  Because various in-painting structs are initialized to the size of the source image, this actually results in a net reduction in memory usage, despite the larger mask space requirement.

Unfortunately, all these geometry changes completely break the algorithm for layers that are not the same size as the parent image.  I will fix this shortly.