Content-aware fill: reduce initialization time by another 20-30%
This is just another standard optimization of switching to insertion sort when a quicksort sub-list is small.  (Plus a few other minor tweaks.)

This and previous commits have now cut initialization time down to ~5% of overall effect time, which is good enough for me.  Further perf improvements will need to come from within the core algorithm.