pdInpaint: rework to operate on single image...
...instead of separate source/destination images.

This is made possible by the previous commit (splitting out the seamless tile code to a separate class).

The primary benefit of a single image instead of split source/destination images is reduced cache pressure; instead of two sets of pixels being accessed by the core loop, now there is only one, which means the single image's data is more likely to be cached when we need it.

Exact consequences of this are challenging to measure on my laptop (throttling due to heat is a problem on these lengthy back-to-back runs) but looks to be in the neighborhood of ~10-15% speed improvement.  All these small gains are adding up to meaningful improvements overall, and I still have some strategies to try - let alone the benefits of adding a UI and letting the user ramp-down extremely high quality settings when they're unnecessary.