Edit > Content-Aware Fill: settings dialog now available
As part of building a settings dialog for the Content-Aware Fill tool, I have also revisited all default values in the inpainter class.  Given PhotoDemon's emphasis on photographs (it's right there in the name!), some defaults needed to be bumped up while others needed to be bumped down.  I think the current set of values is an excellent fit for PD's intended use-case, and the resulting performance of the algorithm is better than ever.

Speaking of performance, I've tweaked even more code in this version and I think the code is now at a fabulous place on the "quality vs speed" spectrum.  Initialization is now so fast that you can't even see it in the UI, and the inpainting process itself is now on-par performance-wise with all major competitors, and significantly better than something like the Content-Aware Fill plugin for Paint.NET (which is based on the same algorithm used here).  This is a big win for a VB6 project, and I'll take those wins whenever I can.

Note that unlike some competing tools, PhotoDemon's algorithm also works on 32-bpp images (with full transparency support), and it also works along image edges/boundaries with no user input.  This allows it to produce excellent results when inpainting something like missing edge chunks in panorama photos taken by your phone.

I now consider the core algorithm more or less "complete".  GIMP provides some additional options (like filling pixels in outside-in order instead of randomly) and I may add similar options to PhotoDemon pending additional testing.