New function for retrieving a null-padded layer copy...
...without actually modifying the source layer.

Many PhotoDemon operations use the concept of a "null-padded layer".  This is a layer whose size is temporarily changed to match the parent image, and all new "padding" pixels filled with transparency.  This is great because I can work on that copy without regard for boundaries (which is often faster and easier than bounds-checking everything) and when I'm done, PhotoDemon can automatically scan the layer and crop it down to its relevant final size.

Historically, I just asked a layer to convert itself to a null-padded layer, then crop itself accordingly I'm done.  But for some new Edit menu features, I can reduce memory usage by asking for a temporary null-padded layer copy, then applying or discarding that temporary copy as necessary.  (For example, if the user cancels the operation.)

I may actually be able to use this to improve memory usage on some other functions... doing that is TBD