Edit > Fill tool is now complete!
Here's how this new tool works.  IF A SELECTION IS ACTIVE:

1) The selection area will be filled with your choice of brush (same brushes available for the paint bucket, text, and other tools - gradient, pattern, etc).
2) You can also apply a new opacity, blend mode, and alpha mode to your fill settings.
3) You can choose to keep existing layer size (like GIMP), change layer size to the union of the existing layer and the selection (like Photoshop), or crop the layer to the current selection.

A live preview is provided for all of these.

IF A SELECTION DOES NOT EXIST, this tool will simply fill the entire active layer with your fill settings.  A live preview also exists for this.

This allows for all of the same fill features as both Photoshop and GIMP, with excellent performance to boot.

Relates to #401