New Edit > Stroke tool
Relates to #401

This tool allows you to stroke the outline of the current selection.  A bunch of options are provided, and performance looks good!

While this first draft works well, I think I can improve stroke quality with some sort of "line refinement" algorithm that performs post-processing clean-up on the marching-squares output (that's how I generate a vector path from the selection mask).  I'm not having luck finding an existing algorithm for this, but I've got an idea how to do it myself, so that's my next mission.  This would produce noticeably better antialiasing, particularly on ellipse shapes.