Edit > Stroke: greatly improve quality with new line simplifier
All PhotoDemon selections work on a selection mask.  This is an image where black indicates unselected pixels, white indicates selected pixels, and gray levels indicates partial selection.

When stroking a selection, we have to take that mask, analyze it, try to reconstruct outline paths as best we can from the bitmap data, then stroke *those* paths and hope the result looks OK.  It's complicated!

This commit adds new vector strategies that try to maximize the quality of the resulting stroke.  It works pretty darn well, and I may be able to improve it further pending additional research.

For now, however, this feature is "ready to go" and provides a very nice feature that is lacking in similar editors (like Paint.NET).