Magic wand selections: if "layer" sampling is active...
...and the user clicks *outside layer boundaries*, create a new selection that selects everything *but* the active layer, regardless of tolerance.

This mirrors identical behavior in Photoshop.  (Photoshop has tons of little selection tool quirks like this that I assume exist for a reason, and I'm trying to mirror them where I can.)

This commit also fixes a bug where changing the active layer didn't update the backing surface on the *next* magic-wand click, and antialiasing has been reinstated for magic-wand selection boundaries (if that toggle is enabled by the user).