Text tools: code cleanup and bug fixes
PhotoDemon's advanced text tool is IMO one of the app's most impressive features.  It provides Photoshop-level typography control despite being written by one sorry bastard on the most text-hostile OS in existence (Windows text rendering for anything but generic text is a debacle.)

Unfortunately, the code that enables this cool feature is pretty complex and tangled.  It needs a few revisions to clean things up, streamline it, and get everything up to modern PD code standards.

This commit is a step in the right direction.  I hope to add some new features to text layers in the coming days, and I want a stronger foundation before doing so.

Among other small fixes, this also fixes a strong annoyance where switching between two text layers did not update the current text boundary rendering on the viewport, *if* a text tool was active.