KD-tree: new "fast construction but slower retrieval" mode
Normally, I use KD-trees for palette matching, specifically for matching a huge set of input colors (a full image) against a small set of output colors (a palette).  In this scenario, it is beneficial to take the time to build a perfectly balanced KD-tree, because subsequent tree traversal will go much faster, and since we have millions of pixels to match, we want the fastest tree possible.

For auto-constructing 3D LUT files from any arbitrary set of photo edits, the opposite problem arises: I need to match some small sample set of pixels (say an 8x8x8 LUT) against all colors in an image (millions!) to find the best match.  In this case, assembling the initial tree becomes the bottleneck for performance, and it's preferable to build the tree as fast as possible, even if that imposes a penalty to later traversal (because we don't have to traverse the tree that many times).

I have tested this mode on PD's regular palette code and the trade-off works exactly as predicted.  Unfortunately, I'm still not sure it'll be fast enough for e.g. 20-megapixel images... but I need to start somewhere.  (If this goes catastrophically badly, I do have some other structures I might try... but I'd rather not write all that code if I can reuse KD-trees mostly as-is.)