New KD-tree implementation with instantaneous teardown
KD-trees are a cool data structure:

https://en.wikipedia.org/wiki/K-d_tree

PhotoDemon uses them all over the place.  Generally, they are used to match pixels (numbering in the millions) against some subset of colors (numbering in the thousands).

Constructing LUTs, however, requires us to solve the opposite problem: matching some subset of LUT colors (numbering in the thousands) against all colors in the image (numbering in the millions).

PD's existing KD-tree implementation worked fine for this, but because the generated trees were so large, COM teardown imposed a huge perf penalty (90+% of LUT generation time was spent just freeing the KD-tree!).

To work around this, I've now written a separate KD-tree optimized against this special use-case.  It uses a backing array to store tree nodes, and is thus basically instantaneous to teardown.  It also has a roughly identical construction cost.

The traversal of such a tree is unfortunately slower, but the teardown benefits far outweigh this cost.  (Roughly speaking, this tree implementation is 2x slower to traverse, but 100000x faster to teardown - so in a real-world example, the overall LUT export process takes some 15 seconds vs 60 seconds before, when accounting for all the different parts of the process.)