New KD-tree class: large traversal perf improvements
Duh, I should have thought this through before I wrote it the way I did - traversal order matters when deciding whether to search sub-trees, and regardless of the backing memory (an array in this case) I still need to traverse it recursively, because a FILO stack like I was using will search sub-trees in a suboptimal order.  Stupid mistake on my part.

Anyway, the idea here is that Export > Color lookup is now extremely fast, only a few seconds even on images with millions of colors.  (In fact, this new class is so fast that I may need to investigate my regular KD-tree to see if I can replace it with this experimental one!)