pdKDTreeArray: add support for tree balancing, alpha channels
I'd really like to ship just one tree class instead of the three I have now, but I have yet to produce a single version with max performance in all cases.  The array-backed version is a clear winner when node counts are large, but for reasons that elude me it's actually slower to traverse when node counts are small.  I'm not sure why.

It should (theoretically) be possible to write a tree traversal routine that doesn't rely on recursion, which would finally produce a "best in all cases" result, but I've tried several times and it's hard to wrap my head around it because I need all results from child nodes before processing the current one - otherwise it works, but it works slowly.  I'll continue to peck away at this as I have time, but for now, I think I'm close to a new stable release and I don't want to derail that chasing down micro-optimizations