Language Editor: new "3rd-party reference translation" tool
One of my biggest challenges is keeping PhotoDemon language files up-to-date after contributors drop out of the project.  I never want to abandon people's localization work, but if I don't speak a language it's incredibly hard to know how to maintain a localization, especially given the number of features added to a typical PD release.

I've already added an integrated online translation tool to the Language Editor, but that can only go so far - especially for menu text like "Copy".  Is that a verb?  A noun?  If it *is* a verb, which tense should I use?  I have no GD idea.

So I've now added a helpful localization feature: an option to load a .po file from other software as an additional translation reference.  When a phrase from PD matches a phrase from that software, PD can now notify you of the 3rd-party translation choice so you can make an informed decision.

For a non-native speaker like me, this is a huge help when trying to figure out how to update a localization.

To make this work, I basically wrote a quick-and-dirty .po parser, and it is indeed quick (and dirty).  For example, GIMP ships with some 5,500+ phrases in a typical .po and PD will load and parse the full file in < 0.1s.  Localized strings are stored in a homebrew hash table and it all works nicely.

I have deliberately written this new tool so that you *cannot* bulk-copy translations from a reference file.  Translations are typically copyright their original authors and I don't want to encourage anyone to steal or use them inappropriately.

So the way PD's tool works is that the 3rd-party translations are only provided as a reference.  To replace an existing PD translations with a 3rd-party one, you must manually review the translations one-by-one and interact with the dialog to swap the text.  This is by design to try and make this feature helpful but ethically okay.

I honestly don't know if anyone else uses PD's language editing tool (most localizers use text editors AFAIK) but as long as I'm writing these features for myself, I may as well make 'em available to others.

With this, I think I have everything I need to get un-maintained localizations up-to-date for a 9.0 release.