Construct rudimentary "action database" at run-time
This database stores action IDs, corresponding human-readable "process names" (e.g. for the action ID "edit_undo" the human-readable process name is "Undo"), and whether a given action can be repeated or faded by the Edit > Repeat and Edit > Fade menus, respectively.

Previously, PD was trying to auto-determine the repeatability or fadeability (?) of a given action by examining its UNDO data type, but this isn't always reliable.  There are weird edge-cases of actions that are repeatable but not fade-able (e.g. "Grow selection").

Relates to #423