Use new action database for Edit > Repeat behavior
Fixes #423

The new run-time "action database" is used to determine Edit > Repeat (and Edit > Fade behavior).  This allows Edit > Repeat to work when a repeatable action is followed by a non-repeatable action (e.g. an Effect followed by creating a different selection).  The repeatable action is now correctly identified and stored, and can be re-used after the non-repeatable action finishes.

Thank you to @YouveGotMeowxy for catching and reporting!