New Ctrl + Shift + Y shortcut for Edit > Repeat
Relates to #423.  Thank you to @YouveGotMeowxy for the suggestion!

Studying a bunch of other software showed no clear hotkey convention for "Repeat last action" behavior.  The closest thing to a pattern in photo editors is that GIMP and Adobe both use Ctrl + F for "last filter"; but we use that shortcut for Search (a shortcut that is standard across a far larger suite of software!).

I eventually settled on `Ctrl + Shift + Y` because it's close to "Redo" (`Ctrl + Y`) and Redo is next to Repeat in the Edit menu.  "Redo" and "Repeat" are similar enough to make the hotkey slightly more memorable.