List hotkey modifiers in Alt + Shift + Ctrl order
I've noticed that non-Microsoft software uses this pattern, and while it deviates from "standard Windows behavior" it has a lovely side-effect.

Because hotkey text is always right-aligned on menu captions, the most common Ctrl+[x] pattern ends up consistently aligned this way, even if Shift and/or Alt modifiers are used.

This not only looks better, it also promotes understanding - e.g. "Shift + Ctrl + C" for Copy Merged becomes an obvious modification of "Ctrl + C" for Copy, and you visually notice this because the "Ctrl + C" portion of both is visually aligned.