Justified text: rudimentary implementation for lines...
...with whitespace characters.  (We can expand the line only in the whitespace locations.)

Relates to #428 .

The canonical best-case implementation of this feature would be the Knuth-Plass linebreaking algorithm:

http://eprg.org/G53DOC/pdfs/knuth-plass-breaking.pdf

I'd love to tackle that someday, but I've yet to fully wrap my brain around it.  Instead, this commit uses a straightforward greedy approach that was easy to slot into PD's existing line-breaking algorithm.

Still TODO:
+ justifying lines that have no whitespace chars
+ variable handling of the last line in a paragraph (the user should really have control over this, like Photoshop provides)