Advanced text tool: finalize justified alignment
Relates to #428

- Lines with whitespace are now justified using whitespace only.
- Lines without whitespace are justified using inter-character spacing.
- A ton of weird edge-cases have been addressed and fixed (seriously, there are a lot of ways for justified text to go horribly wrong)

Next up is a UI for the user to select trailing line justification behavior.  (Photoshop allows the user to select between left/center/right/justified for just the trailing line, as a separate setting - and it's a nice feature.)

Still TODO is handling multi-paragraph text.  Technically the trailing line setting should apply to the trailing line of *each* paragraph in a text layer, but I want to get it working for *just* the last line of *all* paragraphs before I tackle that.