(ICO) icon export: new option for "use merged image as source"
Relates to #453

When exporting an icon file, PhotoDemon previously tried to match each layer to its most appropriate icon frame.  This mirrors similar behavior in GIMP, and it's what allows PD to round-trip icon files correctly.  (e.g. a layer at size 32x32 with 256 colors will be used as the source for a 32x32 8-bit icon frame, if requested)

This behavior can now be overridden in the icon export dialog, with a new "source" toggle to use the merged image as the source for *all* icon frames.  This will be useful when creating a new icon from scratch.

Thank you to @veksha for this idea!