JPEG XL: live previews now available in export dialog
I have also added support for the optional JPEG XL "effort" parameter, which can be used to save faster-but-larger or slower-but-smaller files.  (Just note that at the maximum setting, saving is *extremely* slow - like save time may be measured in minutes, not seconds.)

Almost ready to merge this, I think!