New preference for default Save As filename behavior
Previously, PD always auto-incremented the suggested filename in Save As *if* the current filename already exists.  This serves two purposes: safety (less chance of overwriting by mistake, since it's assumed Save As means you want a new file) and convenience (skipping the likelihood of an additional "are you sure you want to overwrite" prompt from the common dialog).

There is now a toggle for this behavior in the `Tools > Options > Saving` panel, if you want the "standard" behavior.