File > Save As: look for manually entered file extension changes...
...and preferentially use them to determine image format if certain criteria are met.

Relates to #472 .  Thank you to @hi5 for suggesting this improvement.

This behavior is only triggered under the following conditions:

1) File extension modifications are only considered IF the user did NOT interact with the Save As file format dropdown.  (If the user *did* change the dropdown format, the dropdown index is given precedence over any typed changes.)

2) Any manually entered file extension must clearly correspond to a known image format (e.g. changing "image.jpg" to "image.unknown" will still result in saving to JPEG).

3) Manually entered file extensions must correspond to an image format supported by PhotoDemon for export (e.g. changing "image.jpg" to "image.svg" will still save as JPEG, because currently PhotoDemon only supports SVG as an import format).

If all these criteria are met, the typed file extension will be used and a format-specific options dialog will *always* be raised to give the user a chance to confirm settings.  (e.g. opening "image.jpg", clicking `File > Save As` and renaming to "image.png" will always raise a PNG export dialog).

I believe this behavior mimics the best practices of most* other software, and it reduces the chance of a mistaken destination image format when the Save As format dropdown and target file extension do not match.

Thank you to everyone who contributed to the issue discussion on this one, and I apologize that it took me so long to address the issue.