Change JPEG XL support to on-demand
The JPEG XL format remains in constant flux, so rather than shipping it, PD will attempt to download it if/when a user interacts with a JPEG-XL image.

This is the same approach currently used by the AVIF format, and I hope to extend it to HEIF images in the coming days as well.