Create pdPipeSync.cls
This is a (quick and dirty, for now) copy of pdPipeAsync... but synchronous only.

This is useful for third-party libraries that do not provide progress reports and instead are limited to "all-at-once" dumps of data.

Previously I was simply stealing PD's old synchronous ExifTool interface (which is now only used for library version checks) and passing other libraries, but that was sloppy and the code was not well-implemented.  This is a much better starting place, using streams to receive data and much more robust against esoteric piping errors.

The new JPEG XL interface will use this for interfacing with libjxl.