PNG export: always write pHYs tag
Resolves #493.  Thank you to @Charltsing for reporting.

XMP is 22 years old now, but some software still relies on format-specific tags for storing image resolution.  In the case of PNG, this is a pHYs tag.  For software that can't read XMP metadata, PhotoDemon will now write pHYs tags to PNG files, with an identical copy of the resolution data stored in the XMP block.