Add fixes for new lcms2 2.16 behavior w/ COPY_ALPHA flag
In the new 2.16 release, lcms2 changed the way it handles the COPY_ALPHA flag.  When converting from an opaque format to an alpha-supporting format (e.g. RGB to RGBA), the COPY_ALPHA flag can no longer be supplied without lcms2 erroring out.

PhotoDemon uses a shared pathway for color-managing RGB and RGBA data, so the central lcms2 interface needed to be updated to check this state and toggle OFF the COPY_ALPHA flag as necessary.

Fixes #507.  Thank you to @hi5 for catching and reporting!