File > Export > Animation: overhaul completely
Relates to #513 .  Thank you to @hi5 for their thoughtful discussion on how to improve PD's Export menu.

PD's various `File > Export > Animated GIF/JPEG-XL/PNG/WebP` menus are now merged into a single Export > Animation menu.  This menu launches a standard Save dialog where you can choose between those formats, and the dialog has the same auto-detect-format behavior if you manually type in a file extension.

This change was more cumbersome than I expected.  To ensure smooth workflows, PhotoDemon needs to remember export details for a given image (so that, for example, if you load an animated GIF and export it to an animated WebP, when you return to the Export dialog it needs to auto-suggest WebP as the format - not GIF).  It also needs to choose intelligent defaults when it can.  These little nuances add up and lead to a lot of extra code.

Next up: writing a new `Export > Image to file` menu that allows you to export the active image to arbitrary new formats (without modifying image save state).